+++ DanielsMoneyManagerApi/Data/IFundActionsRepository.cs
@@ -6,7 +6,7 @@ namespace DanielsMoneyManagerApi.Data
     {
         void DeleteFundActions(string fundActionIds, int userId);
         List<FundAction> GetFundActions(DateTime fromTime, DateTime toTime, int fundId, int userId);
-        FundAction InsertFundActions(int fundId, float investmentSum, float currentState, string note, DateTime investmentDate);
-        void UpdateFundActions(int fundActionId, int fundId, float investmentSum, float currentState, string note, DateTime investmentDate);
+        FundAction InsertFundActions(int fundId, decimal investmentSum, decimal currentState, string note, DateTime investmentDate);
+        void UpdateFundActions(int fundActionId, int fundId, decimal investmentSum, decimal currentState, string note, DateTime investmentDate);
     }
 }
