+++ DanielsMoneyManagerApi/Dtos/CashActionDto.cs
@@ -4,7 +4,7 @@
     {
         public int cashActionId { get; set; }
         public int categoryId { get; set; }
-        public float sum { get; set; }
+        public decimal sum { get; set; }
         public string description { get; set; }
         public DateTime cashActionTime { get; set; }
         public DateTime insertTime { get; set; }
