+++ DanielsMoneyManagerApi/Dtos/CashActionInsertDto.cs
@@ -3,7 +3,7 @@
     public class CashActionInsertDto
     {
         public int categoryId { get; set; }
-        public float sum { get; set; }
+        public decimal sum { get; set; }
         public string description { get; set; }
         public DateTime cashActionTime { get; set; }
     }
