+++ DanielsMoneyManagerApi/Dtos/FundActionInsertDto.cs
@@ -3,8 +3,8 @@
     public class FundActionInsertDto
     {
         public int fundId { get; set; }
-        public float investmentSum { get; set; }
-        public float currentState { get; set; }
+        public decimal investmentSum { get; set; }
+        public decimal currentState { get; set; }
         public string? note { get; set; }
         public DateTime investmentDate { get; set; }
     }
