+++ DanielsMoneyManagerApi/Dtos/FundActionUpdateDto.cs
@@ -4,8 +4,8 @@
     {
         public int fundActionId { get; set; }
         public int fundId { get; set; }
-        public float investmentSum { get; set; }
-        public float currentState { get; set; }
+        public decimal investmentSum { get; set; }
+        public decimal currentState { get; set; }
         public string? note { get; set; }
         public DateTime investmentDate { get; set; }
     }
