+++ DanielsMoneyManagerApi/Models/CategoryBalance.cs
@@ -3,7 +3,7 @@
     public class CategoryBalance
     {
         public int Category_ID { get; set; }
-        public float Category_Balance { get; set; }
+        public decimal Category_Balance { get; set; }
         public int Cash_Actions_Count { get; set; }
     }
 }
